refactor geo url logic
