+++ src/constants/index.js
@@ -13,4 +13,5 @@ export const BREAKPOINTS = {
   xl: '1200px',
   max: '1680px'
 }
+export const BASE_PROVIDERS_LAMBDA = "/.netlify/functions/providers"
 export const ALL_PROVIDERS_LAMBDA = "/.netlify/functions/providers?location=All"
