+++ src/utils/buildLocationsSet.js
